+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Upcoming
+
+- Configure ESLint plugin import to resolve TS properly
+
 ### 9.0.0
 
 - Add Typescript configuration and upgrade all dependencies